requirements.txt
pyaml
twitter
wget
